requirements.txt
pandas==2.0.3
numpy==1.24.3
mysql-connector-python==8.0.33
sqlalchemy==2.0.20
python-dotenv==1.0.0
pytest==7.4.0
pyyaml==6.0.1
